requirements.txt
tqdm
numpy
scipy
Pillow
imgaug
imageio
Shapely
scikit-image
opencv-python